% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{chrome.create}
\alias{chrome.create}
\title{create selenium chrome
https://cran.r-project.org/web/packages/RSelenium/vignettes/basics.html}
\usage{
chrome.create()
}
\value{

}
\description{
create selenium chrome
https://cran.r-project.org/web/packages/RSelenium/vignettes/basics.html
}
\examples{
chrome.create()
chrome$open()
chrome$navigate("http://www.pchome.com.tw/")
}
