% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cm.library_load}
\alias{cm.library_load}
\title{檢查是否載入這些套件，如果沒有，就載入。}
\usage{
cm.library_load(...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
檢查是否載入這些套件，如果沒有，就載入。
}
\examples{
cm.library_load(magrittr,tidyr,png,jpeg)
}
