% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{f.rbc.add_colname}
\alias{f.rbc.add_colname}
\title{RBC報表皆是用c1:C256去存取資料庫，故使用表格表的表頭去修改表的欄位名稱}
\usage{
f.rbc.add_colname(df, x, y)
}
\arguments{
\item{x}{欄位名稱所在row}

\item{y}{複合欄位}
}
\value{

}
\description{
RBC報表皆是用c1:C256去存取資料庫，故使用表格表的表頭去修改表的欄位名稱
}
