% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{filter.df.date.yqmw_lastday}
\alias{filter.df.date.yqmw_lastday}
\title{篩選yqmw的最後一天}
\usage{
filter.df.date.yqmw_lastday(df, col, yqmw)
}
\arguments{
\item{col}{column}

\item{yqmw}{"y","q","m","w"}
}
\value{

}
\description{
篩選yqmw的最後一天
}
\examples{
df \%>\%
 tidy.df.date.num_to_date(col=date) \%>\%
 filter.df.date.yqmw_lastday(.,col=date,yqmw="w")
}
