% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{knitr.render2rpubs.first}
\alias{knitr.render2rpubs.first}
\title{Initially render RMD file to Rpubs}
\usage{
knitr.render2rpubs.first(filename)
}
\arguments{
\item{filename}{}
}
\value{

}
\description{
Step1: Need to add surrfix to url of Rpubs
Step2: assign variable by ID of result
}
