% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{knitr.render2rpubs.renew}
\alias{knitr.render2rpubs.renew}
\title{Renew render RMD file to Rpubs}
\usage{
knitr.render2rpubs.renew(filename, rpubs_id)
}
\arguments{
\item{rpubs_id}{}
}
\value{

}
\description{
Renew render RMD file to Rpubs
}
