% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{loops.yq.py.stock.finstatement.parse}
\alias{loops.yq.py.stock.finstatement.parse}
\title{迴圈版的py.stock.finstatement.parse}
\usage{
loops.yq.py.stock.finstatement.parse(stockcode, all = F, posfix = F)
}
\arguments{
\item{stockcode, all=T為所有year和quarter, posfix為df的後綴為股票代碼}{}
}
\value{



}
\description{
迴圈版的py.stock.finstatement.parse
}
\examples{
網路取資料並儲存到db，loop(stockcode+year+quarter)

}
