% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pdf2ggpage}
\alias{pdf2ggpage}
\title{pdf2ggpage}
\usage{
pdf2ggpage(path_file, keyword_grepl_format)
}
\arguments{
\item{keyword_grepl_format}{}
}
\value{

}
\description{
pdf2ggpage
}
\examples{
pdf2ggpage('D:\\\\FOMCpresconf20220316.pdf',"inflation|Inflation")
oppdf \%>\% op()
}
