% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{purrr.query.twsestock}
\alias{purrr.query.twsestock}
\title{基於query.twsestock，抓取多月份}
\usage{
purrr.query.twsestock(stockcode, vec_years)
}
\arguments{
\item{vec_years}{}
}
\value{

}
\description{
基於query.twsestock，抓取多月份
}
