% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{py.stock.finstatement.parse}
\alias{py.stock.finstatement.parse}
\title{透過python去抓取單檔股票的財務報表，資產負債表、損益表與現金流量表}
\usage{
py.stock.finstatement.parse(stock, year, quarter)
}
\arguments{
\item{stockcode}{數字格式與字串格式皆可}
}
\value{

}
\description{
網站:https://mops.twse.com.tw/server-java/t164sb01?step=2&SYEAR=2021&file_name=tifrs-fr1-m1-ci-cr-2330-2021Q2.html
}
\examples{
py.stock.finstatement.parse(2317)
}
