% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{query.twsestock}
\alias{query.twsestock}
\title{抓取TWSE單檔股票價格，但僅能抓單月
係參考此文，以R改寫而得 https://medium.com/%E5%B7%A5%E7%A8%8B%E9%9A%A8%E5%AF%AB%E7%AD%86%E8%A8%98/5%E7%A8%AE%E6%8A%93%E5%8F%96%E5%8F%B0%E8%82%A1%E6%AD%B7%E5%8F%B2%E8%82%A1%E5%83%B9%E7%9A%84%E6%96%B9%E6%B3%95-766bf2ed9d6}
\usage{
query.twsestock(stockcode, date_num)
}
\arguments{
\item{date_num}{}
}
\value{

}
\description{
抓取TWSE單檔股票價格，但僅能抓單月
係參考此文，以R改寫而得 https://medium.com/%E5%B7%A5%E7%A8%8B%E9%9A%A8%E5%AF%AB%E7%AD%86%E8%A8%98/5%E7%A8%AE%E6%8A%93%E5%8F%96%E5%8F%B0%E8%82%A1%E6%AD%B7%E5%8F%B2%E8%82%A1%E5%83%B9%E7%9A%84%E6%96%B9%E6%B3%95-766bf2ed9d6
}
