% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{stock.db.finstatement.get}
\alias{stock.db.finstatement.get}
\title{從本地端資料端取得資料}
\usage{
stock.db.finstatement.get(stockcode)
}
\arguments{
\item{stockcode}{}
}
\value{

}
\description{
從本地端資料端取得資料
}
\examples{
stockcode=2317
res = stock.db.finstatement.get(stockcode)
}
