% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{stock.deallist.drop_to_db}
\alias{stock.deallist.drop_to_db}
\title{從dropbox下載deallist}
\usage{
stock.deallist.drop_to_db()
}
\value{

}
\description{
從dropbox下載deallist
}
\examples{
stock.deallist.drop_to_db()
}
