% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{stock.deallist.twse.get}
\alias{stock.deallist.twse.get}
\title{Title 取得多個日期、多個股票的交易明細}
\usage{
stock.deallist.twse.get(dates, stocks)
}
\arguments{
\item{stocks}{}
}
\value{

}
\description{
Title 取得多個日期、多個股票的交易明細
}
\examples{
stock.deallist.twse.get(c(20220110,20220111),2330)
}
