% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{stock.finstatement.analysis.cf}
\alias{stock.finstatement.analysis.cf}
\title{分析現金流量表}
\usage{
stock.finstatement.analysis.cf(df)
}
\arguments{
\item{df}{}
}
\value{
res包含現金流量(營業、投資、自由、融資)、四象限圖、熱力圖。
}
\description{
分析現金流量表
}
