% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{stock.finstatement.parse.download_allfiles}
\alias{stock.finstatement.parse.download_allfiles}
\title{抓取TWSE整份財報。}
\usage{
stock.finstatement.parse.download_allfiles(year, quarter)
}
\arguments{
\item{quarter}{}
}
\value{

}
\description{
https://mops.twse.com.tw/mops/web/t203sb02
}
\examples{
stock.finstatement.parse.download_allfiles(2017,2)
}
