% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{stock.finstatement.report.cf}
\alias{stock.finstatement.report.cf}
\title{清洗產出跨期間比較之現金流量表。}
\usage{
stock.finstatement.report.cf(res, code)
}
\arguments{
\item{res}{來自res = stock.db.finstatement.get(stockcode)}

\item{code}{}
}
\value{

}
\description{
清洗產出跨期間比較之現金流量表。
}
