% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{stock.finstatement.report.pl}
\alias{stock.finstatement.report.pl}
\title{清洗產出跨期間比較之綜合損益表}
\usage{
stock.finstatement.report.pl(res, code = NULL)
}
\arguments{
\item{res}{來自res = stock.db.finstatement.get(stockcode)}

\item{code}{其實就是stockcode}
}
\value{

}
\description{
清洗產出跨期間比較之綜合損益表
}
