% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{stock.finstatement.tidy.download_allfiles}
\alias{stock.finstatement.tidy.download_allfiles}
\title{抓取TWSE網站中全部財務報表至本機端後，去清整出「類別」(tifrs-fr1-m1-ins-ir)清單以加入爬蟲網址}
\usage{
stock.finstatement.tidy.download_allfiles(year, quarter)
}
\arguments{
\item{quarter}{}
}
\value{

}
\description{
"https://mops.twse.com.tw/server-java/t164sb01?step=3&SYEAR=2021&file_name=tifrs-fr1-m1-ins-ir-2823-2021Q3.html"#中壽
"https://mops.twse.com.tw/server-java/t164sb01?step=3&SYEAR=2021&file_name=tifrs-fr1-m1-fh-cr-2882-2021Q3.html"#國泰金
"https://mops.twse.com.tw/server-java/t164sb01?step=3&SYEAR=2021&file_name=tifrs-fr1-m1-ins-cr-5846-2021Q3.html"#國泰人壽
}
\examples{
res = stock.finstatement.tidy.download_allfiles(2021,3)
需在硬碟裡已有下載資料夾。
}
