% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{stock.finstatement.tidy.pl.margin}
\alias{stock.finstatement.tidy.pl.margin}
\title{產生「營業毛利率」、「營業利益率」、「營業淨利率」、「營業純益率」}
\usage{
stock.finstatement.tidy.pl.margin(stockcode)
}
\arguments{
\item{stockcode}{}
}
\value{

}
\description{
產生「營業毛利率」、「營業利益率」、「營業淨利率」、「營業純益率」
}
\examples{
stock.finstatement.tidy.pl.margin(2330)
}
