% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{stock.hash.get.detail}
\alias{stock.hash.get.detail}
\title{載入最詳細的股票(上市、上櫃、興櫃、公開發行)基本資料
https://mops.twse.com.tw/mops/web/t51sb01}
\usage{
stock.hash.get.detail()
}
\value{

}
\description{
載入最詳細的股票(上市、上櫃、興櫃、公開發行)基本資料
https://mops.twse.com.tw/mops/web/t51sb01
}
