% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{stock.hash.get}
\alias{stock.hash.get}
\title{取得股票代號與股票名稱之對照表}
\usage{
stock.hash.get(dates = NULL)
}
\arguments{
\item{dates}{20211224 數字格式}
}
\value{
df
}
\description{
取得股票代號與股票名稱之對照表
}
\examples{
hash.stock = stock.hash.get
}
