% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{stock.ohlc.get}
\alias{stock.ohlc.get}
\title{get OHLC from db}
\usage{
stock.ohlc.get(dates, stocks)
}
\arguments{
\item{dates}{可以空白，空白抓全期}

\item{stocks}{可以空白，空白抓全部股票}
}
\value{

}
\description{
get OHLC from db
}
