% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{stock.tool.code2name}
\alias{stock.tool.code2name}
\title{stockcode to stockname}
\usage{
stock.tool.code2name(code)
}
\arguments{
\item{code}{}
}
\value{

}
\description{
stockcode to stockname
}
\examples{
stock.tool.code2name(c(2330,2317))
stock.tool.code2name(2330)
}
