% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{stock.tool.date.num_to_date}
\alias{stock.tool.date.num_to_date}
\title{Title 排除週六與週日之非交易日。}
\usage{
stock.tool.date.num_to_date(list)
}
\arguments{
\item{list}{}
}
\value{

}
\description{
Title 排除週六與週日之非交易日。
}
\examples{
stock.tool.date.num_to_date(c(20210511:20220111))
}
