% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{stock.tool.name2code}
\alias{stock.tool.name2code}
\title{stockname to stockcode}
\usage{
stock.tool.name2code(name)
}
\arguments{
\item{name}{公司名稱或公司名稱序列}
}
\value{

}
\description{
stockname to stockcode
}
\examples{
stock.tool.name2code(c("永豐實","台積電"))
stock.tool.name2code("台積電")
}
