% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tidy.df.col.seperate_to_unlimit_col}
\alias{tidy.df.col.seperate_to_unlimit_col}
\title{分隔欄位，但不確定要分幾欄，也不指定分後欄位名稱。}
\usage{
tidy.df.col.seperate_to_unlimit_col(df, col, sep)
}
\arguments{
\item{df}{data.frame}

\item{col}{指定column}

\item{sep}{分隔符號}
}
\value{
data.frame
}
\description{
方法來源:https://stackoverflow.com/questions/4350440/split-data-frame-string-column-into-multiple-columns
}
\examples{
path = "D:\\\\USERS\\\\400201\\\\Downloads\\\\tifrs-2021Q3"
list.files(path) \%>\%
 data.frame() \%>\%
 `colnames<-`("file") \%>\%
 tidy.df.col.seperate_to_unlimit_col(col=file,symbol="-")
}
