% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tidy.df.date.num2char}
\alias{tidy.df.date.num2char}
\title{number transform to AC time format}
\usage{
tidy.df.date.num2char(datasets, col)
}
\arguments{
\item{col}{}
}
\value{

}
\description{
number transform to AC time format
}
