% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tidy.df.date.ym_to_ymd.lastday}
\alias{tidy.df.date.ym_to_ymd.lastday}
\title{日期格式ym轉成ymd，其中d是月份的最後一天}
\usage{
tidy.df.date.ym_to_ymd.lastday(df, col)
}
\arguments{
\item{col}{}
}
\value{

}
\description{
日期格式ym轉成ymd，其中d是月份的最後一天
}
