% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tidy.df.date.ymd.ROC_to_CE}
\alias{tidy.df.date.ymd.ROC_to_CE}
\title{ymd:民國年轉西元年, 並以"/"區分}
\usage{
tidy.df.date.ymd.ROC_to_CE(df, col = NULL)
}
\arguments{
\item{col}{}
}
\value{

}
\description{
ymd:民國年轉西元年, 並以"/"區分
}
