% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tidy.list.date.date_to_num}
\alias{tidy.list.date.date_to_num}
\title{Title 日期格式轉成數字格式}
\usage{
tidy.list.date.date_to_num(list)
}
\arguments{
\item{list}{}
}
\value{

}
\description{
Title 日期格式轉成數字格式
}
\examples{
stock.tool.date.num_to_date(c(20210511:20220111)) \%>\% tidy.list.date.date_to_num()
}
