% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tidy.list.date.num_to_date}
\alias{tidy.list.date.num_to_date}
\title{ymd數字向量轉換成日期(date)向量，因為數字是十進位，但日期不是，
所以此轉換會排除不合規的數字轉日期格式。}
\usage{
tidy.list.date.num_to_date(list)
}
\arguments{
\item{num.output}{控制是要輸出數字格式，還是日期格式。}
}
\value{

}
\description{
ymd數字向量轉換成日期(date)向量，因為數字是十進位，但日期不是，
所以此轉換會排除不合規的數字轉日期格式。
}
\examples{
tidy.list.date.num_to_date(c(20200101:20220101))
}
