% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tool.list.date.weekofmonth}
\alias{tool.list.date.weekofmonth}
\title{計算是那當月的那一週}
\usage{
tool.list.date.weekofmonth(dates)
}
\arguments{
\item{dates}{}
}
\value{

}
\description{
計算是那當月的那一週
}
\examples{
dates<- seq(ymd(20210710), today(), by='day')
tool.list.date.weekofmonth(dates)
}
