% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{wd.trans.name_and_date}
\alias{wd.trans.name_and_date}
\title{轉換watchdog慣用的Date和Name為中文名稱}
\usage{
wd.trans.name_and_date(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
轉換watchdog慣用的Date和Name為中文名稱
}
